Fixes GCC ≥14 build errors (closes #115)

+++ src/merger/paraver/opencl_prv_events.c
@@ -233,8 +233,8 @@ void Enable_OpenCL_Operation (unsigned evttype)
 }
 
 int Translate_OpenCL_Operation (unsigned in_evttype, 
-	unsigned long long in_evtvalue, unsigned *out_evttype,
-	unsigned long long *out_evtvalue)
+	UINT64 in_evtvalue, unsigned *out_evttype,
+	UINT64 *out_evtvalue)
 {
 	unsigned u;
 	struct OpenCL_event_presency_label_st *table;
