+++ src/merger/paraver/pthread_prv_events.h
@@ -31,8 +31,8 @@
 #endif
 
 int Translate_pthread_Operation (unsigned in_evttype, 
-	unsigned long long in_evtvalue, unsigned *out_evttype,
-	unsigned long long *out_evtvalue);
+    UINT64 in_evtvalue, unsigned *out_evttype,
+    UINT64 *out_evtvalue);
 
 #if defined(PARALLEL_MERGE)
 void Share_pthread_Operations (void);
